Refactor blog component structure by extracting CardSection.
Extracted the CardSection logic into a separate `CardSectionComponent` for improved modularity and reusability. Updated `BlogComponent` to utilize the new component, ensuring consistent rendering across different screen sizes.
